requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.4
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
mailjet-rest==1.3.4
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyotp==2.9.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
rpds-py==0.13.1
simplejson==3.19.2
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
whitenoise==6.6.0
